Mod: GridSystem integration
