--- Assets/Scripts/Runtime/CH3/Main/Core/GridManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c93d91d3acd872843bcc613618725576
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
