--- Assets/Scripts/Runtime/CH3/Main/Factory/GridObjectFactory.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: c4baac95ba70413e99c00046efb76b93
-timeCreated: 1744625733
