--- Assets/Scripts/Runtime/CH3/Main/Factory/MineralSpawnzone.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a7d8837f44953be42bd52893861e27f8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
