--- Assets/Scripts/Runtime/CH3/Main/Factory/WorldAreaManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: db8ca6bc585e50b48b8630156b78c4ca
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
